🚚 Move server routes and actions into dedicated folder
